+++ CHANGELOG.md
@@ -1,5 +1,17 @@
+# 4.19.2 - 6/24/24
+- [chore] Update reakeys #251
+
+# 4.19.1 - 6/3/24
+- [fix] Fix camera pan while dragging node #248
+
+# 4.19.0 - 5/30/24
+- [feature] Add dolly camera controls #247
+
+# 4.18.1 - 5/23/24
+- [fix] Fix fit view bug after panning and zooming #246
+
 # 4.18.0 - 5/23/24
-- [feature] Add fitNodesInView function
+- [feature] Add fitNodesInView function #244
 
 # 4.17.4 - 5/8/24
 - [chore] upgrade reakeys